feat: implemented footer component
