+++ package.json
@@ -12,9 +12,11 @@
     "@types/node": "^16.3.1",
     "@types/react": "^17.0.14",
     "@types/react-dom": "^17.0.9",
+    "@types/styled-components": "^5.1.11",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "styled-components": "^5.3.0",
     "typescript": "^4.3.5",
     "web-vitals": "^1.1.2"
   },
