--- src/app/components/Footer/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Footer: React.FunctionComponent = () => {
-    return (
-        <React.Fragment>
-            <p>© 2021</p>
-        </React.Fragment>
-    )
-}
-
-export default Footer;
