+++ src/app/components/Layout/Footer/index.tsx
@@ -1,12 +1,14 @@
 import React from "react";
-
-// styles
 import { WrapperFooter, CopyrightText } from "./coponents";
 
-const Footer: React.FunctionComponent = () => (
-    <WrapperFooter>
-      <CopyrightText>© 2021</CopyrightText>
-    </WrapperFooter>
+const Footer: React.FunctionComponent = () => {
+  return (
+    <React.Fragment>
+      <WrapperFooter>
+        <CopyrightText>© 2021</CopyrightText>
+      </WrapperFooter>
+    </React.Fragment>
   );
+};
 
 export default Footer;
