+++ src/app/components/Layout/Header/index.tsx
@@ -1,38 +1,12 @@
 import React from "react";
 
-// compoments
-import Logo from "./Logo";
-import MenuUser from "./MenuUser";
-
-// styles
 import { WrapperHeader } from "./components";
-
-// types
-// import INavLinkData from "./types/INavLinkData";
-
-// data
-// const NavLinksData: Array<INavLinkData> = [
-//   {
-//     label: 'link-1',
-//     path: '#',
-//   },
-//   {
-//     label: 'link-2',
-//     path: '#',
-//   },
-//   {
-//     label: 'link-3',
-//     path: '#',
-//   },
-// ]
 
 const Header: React.FunctionComponent = () => {
   return (
     <React.Fragment>
       <WrapperHeader>
-        <Logo />
-        {/* <NavLinks navLinksData={NavLinksData}/> */}
-        <MenuUser />
+        <h1>Header</h1>
       </WrapperHeader>
     </React.Fragment>
   );
