+++ src/app/components/Layout/index.tsx
@@ -9,7 +9,6 @@ const Layout: React.FunctionComponent = () => {
     return (
         <React.Fragment>
             <Header />
-                    
             <Footer />
         </React.Fragment>
     )
