+++ src/index.tsx
@@ -2,17 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './app';
 import reportWebVitals from './reportWebVitals';
-import { CssBaseline } from '@material-ui/core';
+
 
 ReactDOM.render(
-  <React.StrictMode>
-    <CssBaseline />
+  <React.StrictMode>    
     <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
